requirements.txt
openai>=1.0.0
anthropic>=0.40.0
click>=8.3.1
rich>=14.2.0
tiktoken>=0.12.0
pydantic>=2.12.5
python-dotenv>=1.2.1
platformdirs>=4.5.1
tomli>=2.4.0
httpx>=0.27.0
httpx-sse>=0.4.0
pyfiglet>=0.8.0
ruff